refactor: User, Student 패키지 리팩토링 (#1086)
* chore : JsonNaming 통일 및 주석 서술

* chore : User, Student Model 메소드 정리

* chore : User Repository 정리

* chore : User Service 클래스 분리

* chore : User Controller 정리

* chore : JsonNaming 통일

* chore : StudentService 분리

* chore : StudentUtil 패키지 생성

* chore : StudentApi 설명 변경

* chore : Test 수정

* chore : User, Student 수정하며 생긴 메소드명 변경
