+++ src/main/java/in/koreatech/koin/admin/student/service/AdminStudentService.java
@@ -54,8 +54,8 @@ public class AdminStudentService {
         validateDepartmentValid(adminRequest.major());
         user.update(adminRequest.nickname(), adminRequest.name(),
             adminRequest.phoneNumber(), UserGender.from(adminRequest.gender()));
-        user.updateStudentPassword(passwordEncoder, adminRequest.password());
-        student.update(adminRequest.studentNumber(), adminRequest.major());
+        user.updatePassword(passwordEncoder, adminRequest.password());
+        student.updateInfo(adminRequest.studentNumber(), adminRequest.major());
         adminStudentRepository.save(student);
 
         return AdminStudentUpdateResponse.from(student);
