+++ src/main/java/in/koreatech/koin/domain/student/dto/StudentLoginResponse.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+// SnakeCase X
 public record StudentLoginResponse(
     @Schema(
         description = "Jwt accessToken",
