+++ src/main/java/in/koreatech/koin/domain/user/controller/UserController.java
@@ -6,29 +6,20 @@ import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
 
 import java.net.URI;
 
-import in.koreatech.koin.domain.student.dto.StudentLoginRequest;
-import in.koreatech.koin.domain.student.dto.StudentLoginResponse;
-import in.koreatech.koin.domain.student.dto.StudentRegisterRequest;
-import in.koreatech.koin.domain.student.dto.StudentUpdateRequest;
-import in.koreatech.koin.domain.student.dto.StudentUpdateResponse;
 import in.koreatech.koin.domain.user.dto.*;
+
 import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
 
 import in.koreatech.koin.domain.user.service.UserService;
+import in.koreatech.koin.domain.user.service.UserValidationService;
 import in.koreatech.koin.global.auth.Auth;
-import in.koreatech.koin.global.host.ServerURL;
-import io.swagger.v3.oas.annotations.Hidden;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
@@ -37,14 +28,7 @@ import lombok.RequiredArgsConstructor;
 public class UserController implements UserApi {
 
     private final UserService userService;
-
-    @GetMapping("/user/coop/me")
-    public ResponseEntity<CoopResponse> getCoop(
-        @Auth(permit = COOP) Integer userId
-    ) {
-        CoopResponse coopResponse = userService.getCoop(userId);
-        return ResponseEntity.ok().body(coopResponse);
-    }
+    private final UserValidationService userValidationService;
 
     @PostMapping("/user/login")
     public ResponseEntity<UserLoginResponse> login(
@@ -72,32 +56,6 @@ public class UserController implements UserApi {
             .body(tokenGroupResponse);
     }
 
-    @DeleteMapping("/user")
-    public ResponseEntity<Void> withdraw(
-        @Auth(permit = {STUDENT, OWNER, COOP}) Integer userId
-    ) {
-        userService.withdraw(userId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/user/check/email")
-    public ResponseEntity<Void> checkUserEmailExist(
-        @ModelAttribute(value = "address")
-        @Valid EmailCheckExistsRequest request
-    ) {
-        userService.checkExistsEmail(request);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/user/check/nickname")
-    public ResponseEntity<Void> checkDuplicationOfNickname(
-        @ModelAttribute("nickname")
-        @Valid NicknameCheckExistsRequest request
-    ) {
-        userService.checkUserNickname(request);
-        return ResponseEntity.ok().build();
-    }
-
     @GetMapping("/user/auth")
     public ResponseEntity<AuthResponse> getAuth(
         @Auth(permit = {STUDENT, OWNER, COOP}) Integer userId
@@ -106,12 +64,38 @@ public class UserController implements UserApi {
         return ResponseEntity.ok().body(authResponse);
     }
 
+    @DeleteMapping("/user")
+    public ResponseEntity<Void> withdraw(
+        @Auth(permit = {STUDENT, OWNER, COOP}) Integer userId
+    ) {
+        userService.withdraw(userId);
+        return ResponseEntity.noContent().build();
+    }
+
     @GetMapping("/user/check/login")
     public ResponseEntity<Void> checkLogin(
-            @ParameterObject @ModelAttribute(value = "access_token")
-            @Valid UserAccessTokenRequest request
+        @ParameterObject @ModelAttribute(value = "access_token")
+        @Valid UserAccessTokenRequest request
     ) {
         userService.checkLogin(request.accessToken());
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/user/check/email")
+    public ResponseEntity<Void> checkUserEmailExist(
+        @ModelAttribute(value = "address")
+        @Valid EmailCheckExistsRequest request
+    ) {
+        userValidationService.checkExistsEmail(request);
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/user/check/nickname")
+    public ResponseEntity<Void> checkDuplicationOfNickname(
+        @ModelAttribute("nickname")
+        @Valid NicknameCheckExistsRequest request
+    ) {
+        userValidationService.checkUserNickname(request);
         return ResponseEntity.ok().build();
     }
 
@@ -120,7 +104,16 @@ public class UserController implements UserApi {
         @Valid @RequestBody UserPasswordCheckRequest request,
         @Auth(permit = {STUDENT, OWNER, COOP}) Integer userId
     ) {
-        userService.checkPassword(request, userId);
+        userValidationService.checkPassword(request, userId);
         return ResponseEntity.ok().build();
+    }
+
+    // 영양사로 옮길 예정
+    @GetMapping("/user/coop/me")
+    public ResponseEntity<CoopResponse> getCoop(
+        @Auth(permit = COOP) Integer userId
+    ) {
+        CoopResponse coopResponse = userService.getCoop(userId);
+        return ResponseEntity.ok().body(coopResponse);
     }
 }
