+++ src/main/java/in/koreatech/koin/domain/user/dto/UserTokenRefreshRequest.java
@@ -2,7 +2,6 @@ package in.koreatech.koin.domain.user.dto;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
@@ -13,7 +12,6 @@ import jakarta.validation.constraints.NotNull;
 public record UserTokenRefreshRequest(
     @Schema(description = "refresh_token", example = "eyJhbGciOiJIUzI1NiJ9", requiredMode = REQUIRED)
     @NotNull(message = "refresh_token을 입력해주세요.")
-    @JsonProperty("refresh_token")
     String refreshToken
 ) {
