+++ src/main/java/in/koreatech/koin/domain/user/dto/UserTokenRefreshResponse.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+// SnakeCase X
 public record UserTokenRefreshResponse(
     @Schema(
         description = "Jwt accessToken",
