+++ src/main/java/in/koreatech/koin/domain/user/repository/UserRepository.java
@@ -47,11 +47,6 @@ public interface UserRepository extends Repository<User, Integer> {
             .orElseThrow(() -> UserNotFoundException.withDetail("id: " + id));
     }
 
-    default User getByNickname(String nickname) {
-        return findByNickname(nickname)
-            .orElseThrow(() -> UserNotFoundException.withDetail("nickname: " + nickname));
-    }
-
     default User getByResetToken(String resetToken) {
         return findAllByResetToken(resetToken)
             .orElseThrow(() -> UserNotFoundException.withDetail("resetToken: " + resetToken));
@@ -60,10 +55,6 @@ public interface UserRepository extends Repository<User, Integer> {
     boolean existsByNickname(String nickname);
 
     void delete(User user);
-
-    List<User> findAllByDeviceTokenIsNotNull();
-
-    Optional<User> findByPhoneNumber(String phoneNumber);
 
     List<User> findAllByName(String name);
